requirements.txt
# Core dependencies
tensorflow>=2.13.0
numpy>=1.24.3
pandas>=2.0.3
scikit-learn>=1.3.0

# Telegram bot
python-telegram-bot>=20.3
python-dotenv>=1.0.0

# NLP and text processing
nltk>=3.8.1

# Data visualization and analysis
matplotlib>=3.7.1
seaborn>=0.12.2
tqdm>=4.65.0

# Additional utilities
requests>=2.31.0
joblib>=1.3.0